Expense-tracking-system/requirements.txt
streamlit==1.49.1
fastapi==0.116.1
pydantic==2.11.9
uvicorn==0.35.0
mysql-connector-python==9.4.0
pandas==2.3.2
requests==2.32.5
pytest==8.4.2



